requirements.txt
exceptiongroup==1.3.0
icalendar==6.3.1
iniconfig==2.1.0
numpy==2.0.2
packaging==25.0
pandas==2.3.2
pluggy==1.6.0
Pygments==2.19.2
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
tomli==2.2.1
typing_extensions==4.15.0
tzdata==2025.2
importlib_metadata==8.7.0
